+++ cli.js
@@ -9,9 +9,11 @@ var args = process.argv.slice(2)
 var USAGE = 'Usage:\n' +
             '  ' + pkg.name + '\n' +
             '  ' + pkg.name + ' "QUERIES"\n' +
+            '  ' + pkg.name + ' --json "QUERIES"\n' +
             '  ' + pkg.name + ' --config="path/to/browserlist/file"\n' +
             '  ' + pkg.name + ' --coverage "QUERIES"\n' +
             '  ' + pkg.name + ' --coverage=US "QUERIES"\n' +
+            '  ' + pkg.name + ' --coverage=US,RU,world "QUERIES"\n' +
             '  ' + pkg.name + ' --env="environment name defined in config"\n' +
             '  ' + pkg.name + ' --stats="path/to/browserlist/stats/file"'
 
@@ -34,7 +36,7 @@ if (isArg('--help') || isArg('-h')) {
   var mode = 'browsers'
   var opts = { }
   var queries
-  var country
+  var areas
 
   for (var i = 0; i < args.length; i++) {
     if (args[i][0] !== '-') {
@@ -55,8 +57,14 @@ if (isArg('--help') || isArg('-h')) {
     } else if (name === '--stats' || name === '-s') {
       opts.stats = value
     } else if (name === '--coverage' || name === '-c') {
-      mode = 'coverage'
-      if (value) country = value
+      if (mode !== 'json') mode = 'coverage'
+      if (value) {
+        areas = value.split(',')
+      } else {
+        areas = ['global']
+      }
+    } else if (name === '--json') {
+      mode = 'json'
     } else {
       error('Unknown arguments ' + args[i] + '.\n\n' + USAGE)
     }
@@ -84,29 +92,55 @@ if (isArg('--help') || isArg('-h')) {
     }
   }
 
+  var coverage
   if (mode === 'browsers') {
     browsers.forEach(function (browser) {
       process.stdout.write(browser + '\n')
     })
-  } else {
-    var stats
-    if (country) {
-      stats = country
-    } else if (opts.stats) {
-      stats = JSON.parse(fs.readFileSync(opts.stats))
-    }
-    var result = browserslist.coverage(browsers, stats)
-    var round = Math.round(result * 100) / 100.0
+  } else if (areas) {
+    coverage = areas.map(function (area) {
+      var stats
+      if (area !== 'global') {
+        stats = area
+      } else if (opts.stats) {
+        stats = JSON.parse(fs.readFileSync(opts.stats))
+      }
+      var result = browserslist.coverage(browsers, stats)
+      var round = Math.round(result * 100) / 100.0
 
-    var end = 'globally'
-    if (country && country !== 'global') {
-      end = 'in the ' + country.toUpperCase()
-    } else if (opts.stats) {
-      end = 'in custom statistics'
-    }
+      return [area, round]
+    })
 
-    process.stdout.write(
-      'These browsers account for ' + round + '% of all users ' +
-            end + '\n')
+    if (mode === 'coverage') {
+      var prefix = 'These browsers account for '
+      process.stdout.write(prefix)
+      coverage.forEach(function (data, index) {
+        var area = data[0]
+        var round = data[1]
+        var end = 'globally'
+        if (area && area !== 'global') {
+          end = 'in the ' + area.toUpperCase()
+        } else if (opts.stats) {
+          end = 'in custom statistics'
+        }
+
+        if (index !== 0) {
+          process.stdout.write(prefix.replace(/./g, ' '))
+        }
+
+        process.stdout.write(round + '% of all users ' + end + '\n')
+      })
+    }
+  }
+
+  if (mode === 'json') {
+    var data = { browsers: browsers }
+    if (coverage) {
+      data.coverage = coverage.reduce(function (object, j) {
+        object[j[0]] = j[1]
+        return object
+      }, { })
+    }
+    process.stdout.write(JSON.stringify(data, null, '  ') + '\n')
   }
 }